api/requirements.txt
aio-pika==8.2.5 ; python_version >= "3.10" and python_version < "4"
aiormq==6.4.2 ; python_version >= "3.10" and python_version < "4"
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
asyncpg==0.27.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
dnspython==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.10" and python_version < "4.0"
email-validator==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
fastapi==0.87.0 ; python_version >= "3.10" and python_version < "4.0"
greenlet==2.0.1 ; python_version >= "3.10" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
multidict==6.0.2 ; python_version >= "3.10" and python_version < "4"
pamqp==3.2.1 ; python_version >= "3.10" and python_version < "4"
passlib==1.7.4 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic[dotenv]==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic[email]==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
python-jose==3.3.0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.10" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==1.4.44 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.21.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
uvicorn==0.19.0 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.8.1 ; python_version >= "3.10" and python_version < "4"
